requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9
django-filter==0.11.0
django-toolbelt==0.0.1
djangorestframework==3.3.1
gunicorn==19.4.1
Markdown==2.6.5
psycopg2==2.6.1
static3==0.6.1
